Improve dice support
Refactor dice RNG, avoiding some unnecessary log2() calls, and add a bit
more detail on where it's from and how it works.

Add an estimate of how many dice rolls are expected to generate a
password for a variety of common options.

A further improvement might be to allow the user to give a list of which
dice they have and select the one which requires the fewest average
rolls.